Refactor: 아이콘 위치 및 스토어 사용 변경
